requirements.txt
torch
torchvision
torchaudio
tokenizers
transformers
pandas
numpy
scikit-learn
tensorboard
streamlit
plotly
kagglehub
pyyaml
tqdm